ux: link to new Port for Windows
Yesterday Port added native support, so this points the download button
to the Port latest downloads page
for Windows users.

Also, extract the button into its own component
